web/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.39.0
openai>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.108.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0
flask-sqlalchemy>=3.1.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0
cryptography>=42.0.0
flask-jwt-extended>=4.6.0

# Utilities
python-slugify>=8.0.0
validators>=0.22.0
pyyaml>=6.0.0
